pyproject.toml
[project]
name = "fsedc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arrow>=1.3.0",
    "bs4>=0.0.2",
    "lxml>=5.3.1",
    "pydantic>=2.10.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-sugar>=1.0.0",
    "ruff>=0.8.3",
]
